DoubleSkulls additions to support the format of the SLOBB/ECBBL leagues.
New modules have been added and disabled. 
leaguetables provides a top level menu to quickly get the standard tournament tables for tournaments. This is only really necessary to show the conference module, but I separated it to ensure it works independently. It also shows teams that either have matches scheduled - but not yet played - as well as teams in conferences that don't have matches scheduled yet. 

conference module allows division admins to create conferences within a tournament (via the plugin menu). These provide tournaments the ability to have groups for qualification and show them as a single tournament with the correct ordering within them for the ranking rule.